refactor: migrate src/models/consts.js to typescript
