+++ src/models/calendarObject.js
@@ -7,7 +7,7 @@ import {
 	COMPONENT_NAME_EVENT,
 	COMPONENT_NAME_JOURNAL,
 	COMPONENT_NAME_VTODO,
-} from './consts.js'
+} from './consts.ts'
 
 /**
  * Creates a complete calendar-object-object based on given props
