+++ src/models/calendarShare.js
@@ -7,7 +7,7 @@ import {
 	PRINCIPAL_PREFIX_GROUP,
 	PRINCIPAL_PREFIX_REMOTE_USER,
 	PRINCIPAL_PREFIX_USER,
-} from './consts.js'
+} from './consts.ts'
 
 /**
  * Creates a complete calendar-share-object based on given props
